# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nest API",
    "version": "1.0.0",
    "description": "Generated from the exported Next.js route handlers. Authentication uses the same HttpOnly NextAuth session cookie as the web application; no bearer token is accepted."
  },
  "paths": {
...
    "/api/passkeys": {
      "get": {
        "operationId": "get_api_passkeys",
        "tags": [
          "passkeys"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "secureSessionCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-source-handler": "app/api/passkeys/route.ts"
      },
      "patch": {
        "operationId": "patch_api_passkeys",
        "tags": [
          "passkeys"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "secureSessionCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-source-handler": "app/api/passkeys/route.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_api_passkeys",
        "tags": [
          "passkeys"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "secureSessionCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-source-handler": "app/api/passkeys/route.ts"
      }
    },
...
    "/api/receivables/{id}/close": {
      "post": {
        "operationId": "post_api_receivables_id_close",
        "tags": [
          "receivables"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "secureSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 191
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-source-handler": "app/api/receivables/[id]/close/route.ts",
        "x-request-schemas": [
          "CloseReceivableSchema"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/receivables/budget-summary": {
      "get": {
        "operationId": "get_api_receivables_budget_summary",
        "tags": [
          "receivables"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "secureSessionCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-source-handler": "app/api/receivables/budget-summary/route.ts"
      }
    },
...
    "/api/receivables/summary": {
      "get": {
        "operationId": "get_api_receivables_summary",
        "tags": [
          "receivables"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "secureSessionCookie": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-source-handler": "app/api/receivables/summary/route.ts"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie used in local and non-secure deployments."
      },
      "secureSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-next-auth.session-token",
        "description": "Secure NextAuth session cookie used by HTTPS production deployments."
      },
      "cronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scheduler-only CRON_SECRET. This is not a user API credential."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "issues": {}
        }
      },
...
    },
    "responses": {
      "InvalidRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Unauthenticated": {
        "description": "A valid session is required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The session lacks the required workspace role",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with current state or idempotency history",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "The JSON body or query failed schema validation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A transient dependency is unavailable",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  }
}